+++ CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes to the **kdb VS Code extension** are documented in this file
 
 ### Enhancements
 
-- Refactored extension walkthrough for KDB-X GA
+- Refactored extension welcome page for KDB-X GA
 - Ability to execute SQL files on REPL
 - Removed `local` connection